backend/requirements.txt
pipecat-ai[speechmatics,cartesia,openai,daily]>=0.0.40
fastapi
uvicorn
websockets
python-dotenv
aiohttp
